WIP: dynamic workouts
Instead of having a fixed schedule of which exercises should be done
each day, introduce a new option to "code" which day you should do.

This decision could be based of the measurements of the last workouts,
or in a measurement made that same day after warming up.

Example, if you have a device to measure max strength or speed, you can
do some measures after warming up, introduce them as "measurements" and
let the dynamic workout choose which "day" (set of exercises) you should
be doing.